Move AABB from hitable.rs to aabb.rs
