+++ src/hitable.rs
@@ -4,7 +4,7 @@ use crate::{
         Boxy, ConstantMedium, FlipFace, MovingSphere, RotateY, Sphere, Translate, XYRect, XZRect,
         YZRect,
     },
-    Float, Ray, Vec3,
+    Float, Ray, Vec3, AABB,
 };
 use rand::prelude::*;
 
@@ -219,52 +219,6 @@ impl HitableList {
     // TODO: fixme, silly
     pub fn into_hitable(self) -> Hitable {
         Hitable::HitableList(self)
-    }
-}
-
-#[derive(Clone, Copy)]
-pub struct AABB {
-    pub min: Vec3,
-    pub max: Vec3,
-}
-
-impl AABB {
-    pub fn new(min: Vec3, max: Vec3) -> AABB {
-        AABB { min, max }
-    }
-
-    /// "An Optimized AABB Hit Method" https://raytracing.github.io/books/RayTracingTheNextWeek.html
-    pub fn hit(&self, ray: &Ray, mut tmin: Float, mut tmax: Float) -> bool {
-        for a in 0..3 {
-            let invd = 1.0 / ray.direction[a];
-            let mut t0 = (self.min[a] - ray.origin[a]) * invd;
-            let mut t1 = (self.max[a] - ray.origin[a]) * invd;
-            if invd < 0.0 {
-                std::mem::swap(&mut t0, &mut t1);
-            }
-            tmin = if t0 > tmin { t0 } else { tmin };
-            tmax = if t1 < tmax { t1 } else { tmax };
-            if tmax <= tmin {
-                return false;
-            }
-        }
-        true
-    }
-
-    pub fn surrounding_box(box0: AABB, box1: AABB) -> AABB {
-        let small: Vec3 = Vec3::new(
-            (box0.min.x).min(box1.min.x),
-            (box0.min.y).min(box1.min.y),
-            (box0.min.z).min(box1.min.z),
-        );
-
-        let big: Vec3 = Vec3::new(
-            (box0.max.x).max(box1.max.x),
-            (box0.max.y).max(box1.max.y),
-            (box0.max.z).max(box1.max.z),
-        );
-
-        AABB::new(small, big)
     }
 }
