+++ src/objects/rotate.rs
@@ -1,7 +1,7 @@
 use crate::{
-    hitable::{HitRecord, Hitable, AABB},
+    hitable::{HitRecord, Hitable},
     ray::Ray,
-    Float, Vec3,
+    Float, Vec3, AABB,
 };
 use rand::prelude::*;
 use std::sync::Arc;
